Refactor: main.go を分割し、handlers.go と models.go を作成 #7
